Move form validation to utils
