+++ cmd/test-server/handlers.go
@@ -2,15 +2,12 @@ package main
 
 import (
 	"encoding/json"
-	"errors"
 	"html/template"
 	"net"
 	"net/http"
 	"path"
-	"regexp"
 	"time"
 
-	"github.com/google/uuid"
 	"github.com/gorilla/websocket"
 )
 
@@ -20,26 +17,6 @@ func serveFormTemplate(w http.ResponseWriter) {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
-}
-
-func validateForm(w http.ResponseWriter, r *http.Request) (FormDetails, error) {
-	if err := r.ParseForm(); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return FormDetails{}, nil
-	}
-	if m, _ := regexp.MatchString(`^\w+@brave\.com$`, r.FormValue("email")); !m {
-		return FormDetails{}, errors.New("Invalid Input")
-	}
-	if r.FormValue("exp_type") != "vpn" && r.FormValue("exp_type") != "direct" {
-		return FormDetails{}, errors.New("Invalid Input")
-	}
-	details := FormDetails{
-		UUID:      uuid.NewString(),
-		Timestamp: time.Now().UTC().Format("2006-01-02T15:04:05.000000"),
-		Contact:   r.FormValue("email"),
-		ExpType:   r.FormValue("exp_type"),
-	}
-	return details, nil
 }
 
 func measureHandler(w http.ResponseWriter, r *http.Request) {
